pyproject.toml
[tool.poetry]
name = "jewel-db"
version = "0.1.0"
description = "A modern jewelry inventory system with AI integration"
authors = ["Your Name <your.email@example.com>"]
packages = [
  { include = "jewel_db" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.14"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
sqlmodel = "^0.0.24"
jinja2 = "^3.1.6"
tailwindcss = "*"
pillow = "^11.3.0"
python-dotenv = "^1.1.1"
httpx = "^0.28.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.1"
pre-commit = "^4.2.0"
black = "^25.1.0"
isort = "^6.0.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enforce these rule groups
select = ["E", "F", "W", "C90", "I", "UP"]
# No special ignores (or list only codes your version of ruff knows)
ignore = []
# Match your Python target
target-version = "py311"
# Max line length
line-length = 88
# Folders to skip
exclude = [".git", ".venv", "build", "dist", "__pycache__"]
# File-specific ignores: quoted globs → list of codes
per-file-ignores = { "scripts/*.py" = ["T201"], "__init__.py" = ["F401"] }

[tool.black]
# Black settings
line-length    = 88
target-version = ["py311"]
exclude        = '''
/(
  \.git
 |\.venv
 |build
 |dist
 |__pycache__
)/
'''

[tool.isort]
# Make isort compatible with Black
profile               = "black"
line_length           = 88
multi_line_output     = 3
include_trailing_comma = true
known_first_party     = ["jewel_db", "jewelry_inventory"]
default_section       = "THIRDPARTY"